+++ docs/CHANGELOG.md
@@ -1,15 +1,19 @@
 # Change Log
 
-## [Unreleased] - 2021-05-19
+## [Unreleased] - 2021-05-24
 
  
 ### Added
+
+- `mkref` subcommand. See `celescope rna mkref`, `celescope fusion mkref` for details.
 
 ### Changed
 
 ### Fixed
 
 ### Removed
+
+- Remove `--fusion_pos` from `celescope.fusion.count_fusion`
 
  
 ## [1.2.0] - 2021-05-19
@@ -22,7 +26,7 @@
 
 - Update Seurat from 2.3.4 to 4.0.1.
 
-- `genomeDir` in `celescope.fusion.star_fusion` changed to `fusion_genomeDir` to avoid misunderstanding.
+- `--genomeDir` in `celescope.fusion.star_fusion` changed to `--fusion_genomeDir` to avoid misunderstanding.
 
 - Step `star` sort bam by samtools instead of STAR to avoid potential `not enough memory for BAM sorting` error: https://github.com/alexdobin/STAR/issues/1136